pyproject.toml
[project]
name = "udacity-mnist-classification"
version = "0.1.0"
description = "Udacity MNIST classification project — reusable PyTorch pipeline with config-driven experiments."
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "pandas>=2.1",
    "tqdm>=4.66",
    "tabulate>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "ipykernel>=6.29",
]

[project.scripts]
mnist = "mnist.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mnist"]

# PyTorch wheels: default to CUDA 12.4 on Windows/Linux (matches the desktop with CUDA),
# and fall back to the default PyPI wheels on macOS (which ship with MPS support built in).
[tool.uv.sources]
torch = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
